Promote Github PR ID to int64 (#2446)
Fixes #2443

PR ID and line number are were previously downcasted to int32 in a
number of places despite being mostly treated as an int64 type. While
the likelihood of either of these values exceeding a 32-bit MAX_INT are
minimal, this change makes them 64-bit to err on the side of caution.

+++ internal/controlplane/handlers_githubwebhooks.go
@@ -735,7 +735,7 @@ func getPullRequestInfoFromPayload(
 
 	return &pb.PullRequest{
 		Url:      prUrl,
-		Number:   int32(prNumber),
+		Number:   int64(prNumber),
 		AuthorId: int64(prAuthorId),
 		Action:   action,
 	}, nil
@@ -754,7 +754,7 @@ func reconcilePrWithDb(
 	case "opened", "synchronize":
 		dbPr, err := store.UpsertPullRequest(ctx, db.UpsertPullRequestParams{
 			RepositoryID: dbrepo.ID,
-			PrNumber:     int64(prEvalInfo.Number),
+			PrNumber:     prEvalInfo.Number,
 		})
 		if err != nil {
 			return nil, fmt.Errorf(
@@ -766,7 +766,7 @@ func reconcilePrWithDb(
 	case "closed":
 		err := store.DeletePullRequest(ctx, db.DeletePullRequestParams{
 			RepositoryID: dbrepo.ID,
-			PrNumber:     int64(prEvalInfo.Number),
+			PrNumber:     prEvalInfo.Number,
 		})
 		if err != nil && !errors.Is(err, sql.ErrNoRows) {
 			return nil, fmt.Errorf("cannot delete PR record %d in repo %s/%s",
