+++ internal/engine/ingester/diff/diff.go
@@ -92,7 +92,7 @@ func (di *Diff) Ingest(
 	}
 
 	logger := zerolog.Ctx(ctx).With().
-		Int32("pull-number", pr.Number).
+		Int64("pull-number", pr.Number).
 		Str("repo-owner", pr.RepoOwner).
 		Str("repo-name", pr.RepoName).
 		Logger()
@@ -215,7 +215,8 @@ func ingestFileForFullDiff(filename, patch, patchUrl string) (*pb.PrContents_Fil
 			}
 		} else if strings.HasPrefix(line, "+") {
 			result = append(result, &pb.PrContents_File_Line{
-				Content:    line[1:],
+				Content: line[1:],
+				// see the use of strconv.ParseInt above: this is a safe downcast
 				LineNumber: int32(currentLineNumber),
 			})
